feat(mapper): Publication mapper added.

+++ src/main/java/com/aniDB/aniDB_backend/entity/RecommendAnime.java
@@ -8,7 +8,7 @@ import lombok.*;
 @NoArgsConstructor
 @ToString
 public class RecommendAnime {
-    private Integer memberId;
-    private Integer animeId;
+    private Long memberId;
+    private Long animeId;
     private String discussion;
 }
