+++ src/main/java/com/aniDB/aniDB_backend/entity/RecommendPublication.java
@@ -8,7 +8,7 @@ import lombok.*;
 @NoArgsConstructor
 @ToString
 public class RecommendPublication {
-    private Integer memberId;
-    private Integer publicationId;
+    private Long memberId;
+    private Long publicationId;
     private String discussion;
 }
